Remove old code that opened items directly
Move variables pointing to the background page in shared.js to the returned function, instead of being a module global.
Clean up handling of the modifySelected message.
Move the addTab.execute() to after toggling the recents.
Add MinTabDwellTime to constants.js and use it in background.js and recent-tabs.js.
Change fireAddTab to executeAddTab.